redesign pricing component
