+++ web/public/locales/en.json
@@ -353,6 +353,7 @@
   },
   "main": {
     "seeLiveDemo": "See the live demo",
+    "goToBilling": "Go to billing",
     "faq": {
       "title": "Frequently asked questions",
       "items": [
@@ -1250,17 +1251,16 @@
     "negativePayment": "Negative 'Amount' means that you will have a credit balance of {{currency}}{{dueNowAmount}}. You will be billed {{currency}}0.00 on {{dueNowDate}}. Once the credit balance is used up, your recurring billing amount will be {{currency}}{{nextPaymentAmount}} every billing cycle of your subscription plan.",
     "nextPayment": "Next payment",
     "subscription": "Subscription",
-    "planUsageDesc": "The plan usage shows the number of events (page views, custom events, CAPTCHA completions) that you have used for the current month.",
+    "planUsageDesc": "The plan usage shows the number of events (page views, custom events, error events) that you have used for the current month.",
     "xofy": "{{x}} of {{y}}",
-    "resetDate": "The plan usage resets on the first day of each month.",
+    "resetDate": "The plan usage resets on the first day of each month ({{days}} days until reset)",
     "selectPlan": "Select a subscription plan to unveil full Swetrix potential.",
     "changePlan": "You can change or cancel your subscription plan at any time here.",
-    "membersNotification": "Pageviews, custom events and CAPTCHA completions are counted towards the limit of the subscription of the account owner. People you invite into your projects don't need a subscription.",
+    "membersNotification": "Pageviews, custom events and error events are counted towards the limit of the subscription of the account owner. People you invite into your projects don't need a subscription.",
     "xMonthsFree": "{{amount}} months free",
     "trialDescription": "Your free trial includes access to all features and unlimited data collection limits. When your trial ends, you'll need to select a paid plan that matches your usage to continue using Swetrix.",
     "usageWarningHigh": "You've used {{percentage}}% of your monthly event limit. Consider upgrading your plan to avoid service interruption.",
     "usageWarningCritical": "Warning: You've used {{percentage}}% of your monthly event limit! Upgrade now to ensure uninterrupted service.",
-    "daysUntilReset": "{{days}} days until reset",
     "xPercentUsed": "{{percentage}}% used",
     "xPercentRemaining": "{{percentage}}% remaining"
   },
@@ -1305,34 +1305,25 @@
     "yearlyBilling": "Yearly",
     "switchToMonthly": "Switch to monthly billing",
     "frequency": "Payment frequency",
-    "eventPerMonth": "Events per month",
+    "eventsPerMonth": "events / month",
     "switchToYearly": "Switch to yearly billing",
     "xTrialDaysLeft": "{{amount}} trial days left",
     "trialEnded": "Trial ended",
     "trialEndsToday": "Trial ends today",
     "trialEndsTomorrow": "Trial ends tomorrow",
     "tiers": {
-      "hobby": "Hobby",
-      "freelancer": "Freelancer",
-      "startup": "Startup",
-      "enterprise": "Enterprise",
       "trial": "Free trial",
       "upToXEVMo": "Up to {{amount}} events per month",
-      "xEvMo": "{{amount}} events per month",
       "upToXVMo": "Up to {{amount}} visits per month",
-      "xVMo": "{{amount}} visits per month",
-      "xAlertsSingular": "{{amount}} alerting rule",
-      "xAlertsPlural": "{{amount}} alerting rules",
-      "upToXProjects": "Create up to {{amount}} projects",
       "upToXWebsites": "Add up to {{amount}} websites",
-      "xMoDataRetention": "{{amount}} months of data retention",
       "evXPlanIncl": "Everything {{plan}} plan includes",
-      "smallBusiSupport": "By buying this you are supporting a small business",
       "dataExports": "Unlimited data exports",
       "dataOwnership": "100% data ownership",
-      "dashboards": "Shared & Public Dashboards",
+      "apiAccess": "API access",
       "reports": "Email reports",
-      "userFlowAnalysis": "User flow analysis"
+      "userFlowAnalysis": "User flow analysis",
+      "sessionAnalysis": "Session analysis",
+      "errorTracking": "Error tracking"
     },
     "downgradeTitle": "Are you sure you want to downgrade?",
     "downgradeDesc": "After downgrading, your plan will include fewer projects and available events per month. In case your website contains more traffic than the plan you want to downgrade to includes, this will result in not all events being stored in our database.\nYou will be able to return to your current plan at any time.\nIf you have any questions, you can always email us at {{email}}.",
